rooksafe/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==33.0.0
finnhub-python==2.4.22
fonttools==4.55.0
frozendict==2.4.6
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.3
msgpack==1.1.0
multitasking==0.0.11
numpy==2.1.3
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uwsgi
waitress==3.0.2
webencodings==0.5.1
yfinance==0.2.50
